requirements.txt
Django==3.2.7
Pillow==8.3.2
pytesseract==0.3.6